requirements.txt
# Google Generative AI
google-generativeai

# Hugging Face
langchain-huggingface
transformers
huggingface-hub

# LangChain Core
langchain  # <-- This was missing!
langchain_openai
# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
